Move files into Main folder
